Extract creating raw pipe video to separate method and call it in main method

+++ Views/MainWindow.axaml.cs
@@ -116,11 +116,16 @@ insufficient space or missing folder.";
                 ImageProcessing.CreateFrames(tmpFiles) is IEnumerable<BitmapVideoFrameWrapper> frames)
             {
                 ImageProcessing.ProgressChangedEvent += this.VideoGenerateProgressChanged;
-                var convertedMessage = await ImageProcessing.GenerateVideo(frames, dc.SelectedFramerate, Path.GetDirectoryName(dc.InputFilesList[0].FullName) ?? Path.GetTempPath());
-                ImageProcessing.ProgressChangedEvent -= this.VideoGenerateProgressChanged;
-                if (string.IsNullOrEmpty(convertedMessage)) Directory.Delete(tempFolder, true);
-                if(convertedMessage.Length > 0) dc.ErrorMessage = convertedMessage;
-                dc.VideoProgress = 100;
+                var pipeSource = await ImageProcessing.GenerateVideoPipeSource(frames, dc.SelectedFramerate);
+                if (pipeSource is null) dc.ErrorMessage = "Encountered a problem in generating source from frames";
+                if (pipeSource is not null)
+                {
+                    var convertedMessage = await ImageProcessing.GenerateVideo(pipeSource, Path.GetDirectoryName(dc.InputFilesList[0].FullName) ?? Path.GetTempPath(), frames.Count());
+                    ImageProcessing.ProgressChangedEvent -= this.VideoGenerateProgressChanged;
+                    if (string.IsNullOrEmpty(convertedMessage)) Directory.Delete(tempFolder, true);
+                    if (convertedMessage.Length > 0) dc.ErrorMessage = convertedMessage;
+                    dc.VideoProgress = 100;
+                }
             }
             
             dc.IsVideoConverting = false;
